requirements.txt
django
psycopg2
dj_database_url
djangorestframework